Add a dictionary to keep track of Cryptr objects
If a write operation is in progress and the underlying file system
decides to open another file for reading/writing 'self.cryptr' was
overwritten. This led to incorrect ciphertext in the first write
operation.

To tackle this problem we hold the Cryptr object per file handle until
the file handle is closed.