Use a master key to en-/decrypt files
Only en-/decrypt the master key with the given password. This way we can
change the password later without decrypting and re-encrypting all the
files. We only have to de- and re-encrypt the master key.

We check indirectly if the given password is correct. If not we get an
invalid MAC error from GCM mode.