Open file for writing not only appending
On Linux I had the situation, that a file was open and written to it
several times at offsets % read_size != 0. So when opening we need to
jump at the beginning of the block, decrypt, append the buffer to the
plaintext, encrypt again and write to the beginning of the block. When
opening with appending flag, the new nonce, mac, and ciphertext are
written after the old ones, which leads to a invalid mac next time one
tries to decrypt.